// client/redux/Videos/videos.action-types.ts
export enum ActionTypes {
  FETCH_VIDEOS = 'FETCH_VIDEOS',
  FETCH_VIDEOS_SUCCESS = 'FETCH_VIDEOS_SUCCESS',
  FETCH_VIDEOS_ERROR = 'FETCH_VIDEOS_ERROR',

  LIKE_VIDEOS = 'LIKE_VIDEOS',
  LIKE_VIDEOS_SUCCESS = 'LIKE_VIDEOS_SUCCESS',
  LIKE_VIDEOS_ERROR = 'LIKE_VIDEOS_ERROR',

  DISLIKE_VIDEOS = 'DISLIKE_VIDEOS',
  DISLIKE_VIDEOS_SUCCESS = 'DISLIKE_VIDEOS_SUCCESS',
  DISLIKE_VIDEOS_ERROR = 'DISLIKE_VIDEOS_ERROR',
}